Переписывает функцию отрисовки точек маршрута в дни
