+++ src/components/events-list.js
@@ -1,6 +1,5 @@
 import {createElement} from "../utils";
 
-
 const createEventsListTemplate = () => `<ul class="trip-events__list"></ul>`;
 
 export default class EventsListComponent {
